refactor(texts): name folder service to services
